% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeBalance}
\alias{computeBalance}
\title{Compute Balances from a Item Shares}
\usage{
computeBalance(data_table, multipliers = list(`item:pos` = 1, `item:eq` = 0,
  `item:neg` = -1))
}
\arguments{
\item{data_table}{a data.table in wide format containing item}

\item{multipliers}{list containing multipliers of items, assigned by item and column names}
}
\description{
This function computes balances (i.e. positive - negative items), from item shares
stored in a wide format data.table.
}
\author{
Matthias Bannert, Gabriel Bucu
}

