% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeShares}
\alias{computeShares}
\title{Compute Weighted Shares By Group}
\usage{
computeShares(data_table, variable, weight, by, wide = T)
}
\arguments{
\item{data_table}{a data.table}

\item{variable}{character name of the variable to focus on. The variable must be in the data.table}

\item{weight}{character name of the data.table column that contains a weight.}

\item{by}{character vector of the columns to group by}

\item{wide}{logical if true the result is returned in wide format dcast.}
}
\description{
This function computes weighted shares from a data.table.
computeShares is performance optimized and designed to
work well in bulk operations. The function returns a data.table.
}
\author{
Matthias Bannert, Gabriel Bucur, Oliver Mueller
}

