% Generated by roxygen2 (4.0.2): do not edit by hand
\name{computeWeightedMean}
\alias{computeWeightedMean}
\title{Compute Weighted Mean by Group}
\usage{
computeWeightedMean(data_table, variable, weight, by)
}
\arguments{
\item{data_table}{a data.table}

\item{variable}{character name of the variable to focus on. The variable must be in the data.table}

\item{weight}{character name of the data.table column that contains a weight.}

\item{by}{character vector of the columns to group by}
}
\description{
This function computes the weighted mean of variable groups from a data.table.
computeWeightedMean is performance optimized and designed to
work well in bulk operations. The function returns a data.table.
}
\examples{
# load library and dataset
library(panelaggregation)
data(btsdemo)
head(btsdemo)
# adapt the levels to positive, equal and negative
# in order to suit the naming defaults. other levels work too, 
# but you'd need to specify multipliers in computeBalance then
levels(btsdemo$question_1) <- c("pos","eq","neg")

# compute the weighted shares and display store in wide format 
# to get a basis for further steps
level1 <- computeShares(btsdemo,"question_1","weight",by = c("date_qtrly","group"))

# compute balance, don't have to do much here, because
# 3 items is the default
level1_wbalance <- computeBalance(level1)

# Select a particular grouping combination and a timeseries that 
# should be extracted from the data_table
ts1 <- extractTimeSeries(level1_wbalance,
                  "date_qtrly",
                  list(group = "C"),
                  freq = 4,
                  item = "balance",
                  variable = "question_1")
ts1
# Plot a standard R ts using the plot method for ts
plot(ts1,main = attributes(ts1)$ts_key)

# DON'T RUN
# Optionally you can store the series into a time series db using the
# corresponding package... 
# Note that you can use the attribute to specify a timeseries key dynamically ! 
# library(timeseriesdb)
# con <- dbConnect(PostgreSQL(),user = "yourname",dbname="yourdb",host = "yourhost")
# storeTimeseries("ts1",attributes(ts1)$ts_key,"con")
}
\author{
Matthias Bannert, Gabriel Bucur
}

