% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doUniqueCJ.R
\name{doUniqueCJ}
\alias{doUniqueCJ}
\title{Performs a Cross Join of Unique combinations}
\usage{
doUniqueCJ(dt, cols)
}
\arguments{
\item{dt}{data.table}

\item{cols}{character vector that denotes names of relevant columns}
}
\description{
This function makes use of \code{\link{CJ}} function of the data.table package to perform a
cross join. The function makes sure that the combinations are unique and removes NAs before
joining. doUniqueCJ is rather not used as a standalone function but inside \code{\link{computeShares}}.
}
\author{
Matthias Bannert, Gabriel Bucur
}

