% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractTimeSeries}
\alias{extractTimeSeries}
\title{Extract a Time Series from a Data.table}
\usage{
extractTimeSeries(data_table, time_column, group_list, freq, variable)
}
\arguments{
\item{data_table}{a data.table}

\item{time_column}{character name of the column which contains the time index}

\item{group_list}{list}

\item{freq}{integer value either 4 denoting quarterly frequency or 12 denoting quarterly frequency}

\item{variable}{character name of the column which contains the item that is extracted from the data.table}
}
\description{
This functions extracts time series from data.table columns and returns object of class ts.
}
\author{
Matthias Bannert, Gabriel Bucur
}

