% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinDataTables.R
\name{joinDataTables}
\alias{joinDataTables}
\title{Joins two data.tables based on keys}
\usage{
joinDataTables(dt_1, dt_2, key_1, key_2 = key_1)
}
\arguments{
\item{dt_1}{first data.table}

\item{dt_2}{second data.table}

\item{key_1}{character vector of key columns for first data.table}

\item{key_2}{character vector of key columns for second data.table}
}
\value{
joined data.table
}
\description{
This function joins two data.table objects, given a common key, which can have different names in the two tables.
In the latter case, the sequence of the names is crucial. Make sure that the key columns match exactly.
}
\author{
Matthias Bannert, Gabriel Bucur
}

