% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeBalance.R
\name{computeBalance}
\alias{computeBalance}
\title{Compute Balances from a Item Shares}
\usage{
computeBalance(data_table, multipliers = list(item_pos = 1, item_eq = 0,
  item_neg = -1))
}
\arguments{
\item{data_table}{a data.table in wide format containing item}

\item{multipliers}{list containing multipliers of items, assigned by item and column names}
}
\description{
This function computes balances (i.e. positive - negative items), from item shares
stored in a wide format data.table.
}
\author{
Matthias Bannert, Gabriel Bucu
}

